pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "mysql-connector-python>=9.5.0",
    "pillow>=12.0.0",
]
